+++ src/OpenTelemetry/Metrics/Export/MetricData.cs
@@ -24,6 +24,8 @@ namespace OpenTelemetry.Metrics.Export
     /// </summary>
     public abstract class MetricData
     {
+        public DateTime StartTimestamp { get; set; }
+
         public DateTime Timestamp { get; set; }
 
         public IEnumerable<KeyValuePair<string, string>> Labels { get; set; }