+++ src/OpenTelemetry.Exporter.Console/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 0.8.0-beta.1
+
+Released 2020-Nov-5
+
+* Add extension method to add `ConsoleExporter` for logs
+  ([#1452](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1452))
+
+* Generalized `ConsoleExporter` to add support for logs
+  ([#1438](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1438))
+
 ## 0.7.0-beta.1
 
 Released 2020-Oct-16