+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,12 +2,19 @@
 
 ## Unreleased
 
+## 0.8.0-beta.1
+
+Released 2020-Nov-5
+
 * Renamed TextMapPropagator to TraceContextPropagator, CompositePropapagor
   to CompositeTextMapPropagator. IPropagator is renamed to TextMapPropagator
   and changed from interface to abstract class.
   ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1427))
-* Propagators.DefaultTextMapPropagator will be used as the default Propagator
+* Propagators.DefaultTextMapPropagator will be used as the default Propagator.
   ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1428))
+* Removed Propagator from Instrumentation Options. Instrumentation now always
+  respect the Propagator.DefaultTextMapPropagator.
+  ([#1448](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1448))
 
 ## 0.7.0-beta.1